backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Marcus Schiesser <mail@marcusschiesser.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11,<3.12"
fastapi = "^0.104.1"
uvicorn = { extras = ["standard"], version = "^0.23.2" }
python-dotenv = "^1.0.0"
unstructured = "0.10.30"
Jinja2 = "3.1.2"
llama-index-core = "^0.10.18.post1"
llama-index-embeddings-openai = "^0.1.6"
llama-index-llms-openai = "^0.1.7"
llama-index-readers-file = "^0.1.8"
llama-index-agent-openai = "^0.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
